+++ common-test/src/main/java/io/customer/commontest/BaseTest.kt
@@ -7,13 +7,11 @@ import androidx.test.platform.app.InstrumentationRegistry
 import io.customer.commontest.util.DispatchersProviderStub
 import io.customer.sdk.CustomerIOConfig
 import io.customer.sdk.data.model.Region
+import io.customer.sdk.data.store.Client
 import io.customer.sdk.data.store.DeviceStore
 import io.customer.sdk.di.CustomerIOComponent
-import io.customer.sdk.util.CioLogLevel
-import io.customer.sdk.util.DateUtil
-import io.customer.sdk.util.DispatchersProvider
-import io.customer.sdk.util.JsonAdapter
-import io.customer.sdk.util.Seconds
+import io.customer.sdk.module.CustomerIOModuleConfig
+import io.customer.sdk.util.*
 import okhttp3.ResponseBody.Companion.toResponseBody
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
@@ -37,7 +35,7 @@ abstract class BaseTest {
 
     protected lateinit var cioConfig: CustomerIOConfig
 
-    protected val deviceStore: DeviceStore = DeviceStoreStub().deviceStore
+    protected lateinit var deviceStore: DeviceStore
     protected lateinit var dispatchersProviderStub: DispatchersProviderStub
 
     protected lateinit var di: CustomerIOComponent
@@ -51,9 +49,43 @@ abstract class BaseTest {
     protected lateinit var mockWebServer: MockWebServer
     protected lateinit var dateUtilStub: DateUtilStub
 
+    protected fun createConfig(
+        client: Client = Client.Android(sdkVersion = "1.0.0-alpha.6"),
+        siteId: String = this.siteId,
+        apiKey: String = "xyz",
+        region: Region = Region.EU,
+        timeout: Long = 100,
+        autoTrackScreenViews: Boolean = true,
+        autoTrackDeviceAttributes: Boolean = true,
+        backgroundQueueMinNumberOfTasks: Int = 10,
+        backgroundQueueSecondsDelay: Double = 30.0,
+        backgroundQueueTaskExpiredSeconds: Double = Seconds.fromDays(3).value,
+        logLevel: CioLogLevel = CioLogLevel.DEBUG,
+        trackingApiUrl: String? = null,
+        targetSdkVersion: Int = android.os.Build.VERSION_CODES.R,
+        configurations: Map<String, CustomerIOModuleConfig> = emptyMap()
+    ) = CustomerIOConfig(
+        client = client,
+        siteId = siteId,
+        apiKey = apiKey,
+        region = region,
+        timeout = timeout,
+        autoTrackScreenViews = autoTrackScreenViews,
+        autoTrackDeviceAttributes = autoTrackDeviceAttributes,
+        backgroundQueueMinNumberOfTasks = backgroundQueueMinNumberOfTasks,
+        backgroundQueueSecondsDelay = backgroundQueueSecondsDelay,
+        backgroundQueueTaskExpiredSeconds = backgroundQueueTaskExpiredSeconds,
+        logLevel = logLevel,
+        trackingApiUrl = trackingApiUrl,
+        targetSdkVersion = targetSdkVersion,
+        configurations = configurations
+    )
+
+    protected open fun setupConfig() = createConfig()
+
     @Before
     open fun setup() {
-        cioConfig = CustomerIOConfig(siteId, "xyz", Region.EU, 100, null, true, true, 10, 30.0, Seconds.fromDays(3).value, CioLogLevel.DEBUG, null)
+        cioConfig = setupConfig()
 
         // Initialize the mock web server before constructing DI graph as dependencies may require information such as hostname.
         mockWebServer = MockWebServer().apply {
@@ -74,6 +106,7 @@ abstract class BaseTest {
         dateUtilStub = DateUtilStub().also {
             di.overrideDependency(DateUtil::class.java, it)
         }
+        deviceStore = DeviceStoreStub().getDeviceStore(cioConfig)
         dispatchersProviderStub = DispatchersProviderStub().also {
             di.overrideDependency(DispatchersProvider::class.java, it)
         }