+++ messagingpush/src/main/java/io/customer/messagingpush/util/PushTrackingUtil.kt
@@ -1,4 +1,4 @@
-package io.customer.sdk.util
+package io.customer.messagingpush.util
 
 import android.os.Bundle
 import io.customer.sdk.data.request.MetricEvent
@@ -6,20 +6,20 @@ import io.customer.sdk.repository.TrackRepository
 
 interface PushTrackingUtil {
     fun parseLaunchedActivityForTracking(bundle: Bundle): Boolean
+
+    companion object {
+        const val DELIVERY_ID_KEY = "CIO-Delivery-ID"
+        const val DELIVERY_TOKEN_KEY = "CIO-Delivery-Token"
+    }
 }
 
 class PushTrackingUtilImpl(
     private val trackRepository: TrackRepository
 ) : PushTrackingUtil {
 
-    companion object {
-        const val DELIVERY_ID_KEY = "CIO-Delivery-ID"
-        const val DELIVERY_TOKEN_KEY = "CIO-Delivery-Token"
-    }
-
     override fun parseLaunchedActivityForTracking(bundle: Bundle): Boolean {
-        val deliveryId = bundle.getString(DELIVERY_ID_KEY)
-        val deliveryToken = bundle.getString(DELIVERY_TOKEN_KEY)
+        val deliveryId = bundle.getString(PushTrackingUtil.DELIVERY_ID_KEY)
+        val deliveryToken = bundle.getString(PushTrackingUtil.DELIVERY_TOKEN_KEY)
 
         if (deliveryId == null || deliveryToken == null) return false